Rename RotationFirstDualQuaternion
