+++ s2e-ff/src/library/math/rotation_first_dual_quaternion.hpp
